Continue with init module; modularize and organize miscellaneous code
